Added coordinates to result

+++ Constants.cs
@@ -2,42 +2,42 @@
 
 internal class Constants
 {
-    public static int[,] Piece1 = {
+    public static byte[,] Piece1 = {
         { 1, 1, 1 },
         { 0, 1, 0 },
         { 0, 1, 0 } };
 
-    public static int[,] Piece2 = {
+    public static byte[,] Piece2 = {
         {1, 1, 1 },
         {1, 0, 1}};
 
-    public static int[,] Piece3 = {
+    public static byte[,] Piece3 = {
         {1, 1},
         {1, 1},
         {1, 1}};
 
-    public static int[,] Piece4 = {
+    public static byte[,] Piece4 = {
         {1, 1, 1, 0},
         {0, 0, 1, 1}};
 
-    public static int[,] Piece5 = {
+    public static byte[,] Piece5 = {
         {1, 1, 1, 1},
         {1, 0, 0, 0}};
 
-    public static int[,] Piece6 = {
+    public static byte[,] Piece6 = {
         {1, 1, 1, 1},
         {0, 1, 0, 0}};
 
-    public static int[,] Piece7 = {
+    public static byte[,] Piece7 = {
         {1, 1, 1},
         {1, 1, 0}};
 
-    public static int[,] Piece8 = {
+    public static byte[,] Piece8 = {
         {1, 1, 1},
         {1, 0, 0},
         {1, 0, 0}};
 
-    public static int[,] Board = {
+    public static byte[,] Board = {
         {0, 0, 0, 0, 0, 0, 1},
         {0, 0, 0, 0, 0, 0, 1},
         {0, 0, 0, 0, 0, 0, 0},
